+++ src/main/java/com/LilliputSalon/SalonApp/repository/AppointmentHistoryRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.LilliputSalon.SalonApp.domain.AppointmentHistory;
 
-public interface AppointmentHistoryRepository extends JpaRepository<AppointmentHistory, Integer> {
+public interface AppointmentHistoryRepository extends JpaRepository<AppointmentHistory, Long> {
 
-    List<AppointmentHistory> findByAppointmentId(Integer appointmentId);
+    List<AppointmentHistory> findByAppointmentId(Long appointmentId);
 }
