+++ src/main/java/com/LilliputSalon/SalonApp/repository/BreakTimeRepository.java
@@ -8,10 +8,10 @@ import org.springframework.stereotype.Repository;
 import com.LilliputSalon.SalonApp.domain.BreakTime;
 
 @Repository
-public interface BreakTimeRepository extends JpaRepository<BreakTime, Integer> {
+public interface BreakTimeRepository extends JpaRepository<BreakTime, Long> {
 
     // valid — find all breaks for one availability block
-    List<BreakTime> findByAvailability_AvailabilityId(Integer availabilityId);
+    List<BreakTime> findByAvailability_AvailabilityId(Long availabilityId);
 
     List<BreakTime> findByAvailability_User_Id(Long userId);
